+++ src/components/NewComps/Banner/StackedBanner/index.js
@@ -15,28 +15,28 @@ export default function StackedBanner({ compData }) {
           />
         ) : null}
 
-        <h1 className="text-center c-fw-400 heading garmond-font col-primary">
+        <h1 className="text-center fontw-400 heading garmond-font font-primary">
           {compData?.heading}
         </h1>
         <div className="col-lg-8 mx-auto">
-          <p className="text-center col-grey-deep">{compData?.subHeading}</p>
+          <p className="text-center font-grey-deep">{compData?.subHeading}</p>
         </div>
       </div>
       <div className="pb-5 container d-flex gap-2 gap-md-4 justify-content-center align-items-center flex-column flex-md-row">
-        {compData?.buttons?.length > 0 && (
-          <a href={compData?.buttons[0].link} target="_blank">
-            <button className="btn btn-primary">
-              {compData?.buttons[0].text}
+        {compData?.buttons?.map((button, index) => (
+          <a
+            key={index}
+            href={button.link}
+            target="_blank"
+          >
+            <button
+              className={`btn ${index === 0 ? "btn-primary" : "btn-primary-outline"
+                }`}
+            >
+              {button.text}
             </button>
           </a>
-        )}
-        {compData?.buttons?.length > 1 && (
-          <a href={compData?.buttons[1].link} target="_blank">
-            <button className="btn btn-primary-outline">
-              {compData?.buttons[1].text}
-            </button>
-          </a>
-        )}
+        ))}
       </div>
     </section>
   );
