+++ src/components/NewComps/Features/GridFeatures/GridFeatures.module.scss
@@ -1,0 +1,4 @@
+.card{
+    min-height: 330px;
+    background: radial-gradient(50% 50% at 50% 50%, var(--col-accent-light) 0%, var(--col-white) 100%);
+}
