+++ src/components/Pages/CloudAccountingSoftware.js
@@ -1,0 +1,19 @@
+import StackedBanner from "../NewComps/Banner/StackedBanner";
+import AlternateFeatures from "../NewComps/Features/AlternateFeatures";
+import GridFeatures from "../NewComps/Features/GridFeatures";
+import FeedbackCards from "../NewComps/FeedbackCards";
+import CTA from "../NewComps/CTA";
+import FAQs from "../NewComps/FAQs";
+
+export default function CloudAccountingSoftware({ pageData, pageInfo }) {
+  return (
+    <>
+      <StackedBanner compData={pageData?.stackedBanner} />
+      <GridFeatures compData={pageData?.characteristics} />
+      <AlternateFeatures features={pageData?.features} />
+      <FeedbackCards testimonials={pageData?.testimonials} />
+      <FAQs faqs={pageData?.faqs} />
+      <CTA compData={pageData?.cta} />
+    </>
+  );
+}
