+++ src/scss/_import.scss
@@ -1,6 +1,7 @@
 @import 'reset';
 @import 'text';
 @import 'var';
+@import 'mixins';
 
 // SCSS Components
 @import 'scsscomps/navbar';
