+++ src/scss/_mixins.scss
@@ -1,0 +1,96 @@
+//====================================UTILITY MIXINS========================================
+
+// Universal mixin for generating utility classes
+@mixin generate-utility-classes(
+  $prefix,
+  $property,
+  $values,
+  $important: false
+) {
+  @each $key, $var in $values {
+    .#{$prefix}-#{$key} {
+      @if $important {
+        #{$property}: var(#{$var}) !important;
+      } @else {
+        #{$property}: var(#{$var});
+      }
+    }
+  }
+}
+
+// Mixin for generating higher specificity border color classes
+@mixin generate-border-classes($color-map) {
+  @each $key, $var in $color-map {
+    .border.border-color-#{$key} {
+      border-color: var(#{$var}) !important;
+    }
+  }
+}
+
+//====================================UTILITY CLASSES========================================
+
+// Define color values map (reusable for both background and text colors)
+$color-values: (
+  primary: --col-primary,
+  dark: --col-dark,
+  grey: --col-grey,
+  grey-faded: --col-grey-faded,
+  dark-grey: --dark-grey,
+  dark-light: --col-dark-light,
+  light: --col-light,
+  white: --col-white,
+  blue: --col-blue,
+  success: --col-success,
+  danger: --col-danger,
+  grey-deep: --col-grey-deep,
+  light-gray: --col-light-gray,
+  light-green: --col-light-green,
+  baby-pink: --col-baby-pink,
+  light-blue: --col-light-blue,
+  soft-blue: --col-soft-blue,
+  faded-blue: --col-faded-blue,
+  light-golden: --col-light-golden,
+  extra-light-blue: --col-extra-light-blue,
+  slate-grey: --col-slate-grey,
+);
+
+// Generate background color classes using universal mixin
+@include generate-utility-classes("bg", "background-color", $color-values);
+
+// Generate text color classes using universal mixin
+@include generate-utility-classes("font", "color", $color-values);
+
+// Generate border color override classes with higher specificity
+@include generate-border-classes($color-values);
+
+// Mixin for generating font-size classes
+@mixin font-size-class($size-var, $important: false) {
+  @if $important {
+    font-size: var(#{$size-var}) !important;
+  } @else {
+    font-size: var(#{$size-var});
+  }
+}
+
+// Mixin for generating multiple font-size classes
+@mixin generate-font-size-classes($prefix, $sizes, $important: false) {
+  @each $key, $var in $sizes {
+    .#{$prefix}-#{$key} {
+      @include font-size-class($var, $important);
+    }
+  }
+}
+
+// Generate font classes using mixin
+@include generate-font-size-classes(
+  "font",
+  (
+    xs: --font-xs,
+    sm: --font-sm,
+    rg: --font-rg,
+    md: --font-md,
+    lg: --font-lg,
+    xl: --font-xl,
+    xxl: --font-xxl,
+  )
+);
