+++ src/scss/_text.scss
@@ -152,3 +152,12 @@
 .font-sub-heading {
   font-size: var(--font-sub-heading);
 }
+
+// font weight
+
+.fontw-400 {
+  font-weight: var(--c-fw-400) !important;
+}
+.fontw-600 {
+  font-weight: var(--c-fw-600) !important;
+}
