+++ src/scss/_var.scss
@@ -34,6 +34,8 @@
   --col-faded-blue: #e9ecff;
   --col-light-golden: #f9dcae;
   --col-extra-light-blue: #f5f6ff;
+  --col-slate-grey:#3F4346;
+  --col-accent-light: #E2E6FF;
 
   //Content Width
   --w-max-content: 900px;
